🐛 fix : Fix database migration

+++ database/migrations/2023_05_06_074457_create_user_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('id')->unique();
+            $table->integer('id')->unique();
             $table->string('email');
             $table->string('password');
             $table->string('nickname');
